pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atv-paperboard"
version = "0.1.2"
description = "Cross-harness HTML artifact toolkit for AI coding agents. Native plugins for Claude Code, Codex CLI, and GitHub Copilot CLI + GitHub Actions recipe for the Copilot Coding Agent. Enforce, render, persist, compound."
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "All The Vibes" }]
requires-python = ">=3.10"
keywords = [
    "ai-agents",
    "ai-coding",
    "llm-tools",
    "claude-code",
    "codex-cli",
    "copilot-cli",
    "github-copilot",
    "copilot-coding-agent",
    "agent-plugin",
    "agent-skills",
    "agent-hooks",
    "cross-harness",
    "design-md",
    "design-system",
    "html-generator",
    "html-artifact",
    "dashboard",
    "cli-tool",
    "developer-tools",
    "python",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[project.scripts]
paperboard = "core.cli:main"

[project.urls]
Homepage = "https://github.com/All-The-Vibes/ATV-PaperBoard"
Repository = "https://github.com/All-The-Vibes/ATV-PaperBoard"
Issues = "https://github.com/All-The-Vibes/ATV-PaperBoard/issues"
Changelog = "https://github.com/All-The-Vibes/ATV-PaperBoard/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
include = ["core*"]
exclude = ["tests*", "teaser*"]

[tool.setuptools.package-data]
core = [
    "templates/**/*.j2",
    "templates/**/*.html",
    "designs/**/*.DESIGN.md",
    "designs/**/.gitkeep",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
markers = [
    "phase0: Phase 0 empirical verification tests (require Node + @google/design.md)",
    "harness_claude_code: Tests requiring a real Claude Code session",
    "harness_codex: Tests requiring a real Codex CLI session",
    "slow: Tests that take >5 seconds",
]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
    # Generated by `python adapters/<harness>/build.py`; mirrors core/
    # source which is already linted at the top level. Linting these
    # would double-report every finding.
    "adapters/*/_dist",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = [
    "E501",  # line length handled by formatter
    # The codebase intentionally uses single-line guard statements
    # (e.g. `if num: head_parts.append(...)`) for readability in tight
    # rendering blocks. Don't flag them.
    "E701",
    "E702",
]
